allow IDs with fractional digits; e.g. 123.1 to represent multiple tags with the same ID in the same project

+++ find_tags_common.hpp
@@ -26,9 +26,22 @@ typedef int Nominal_Frequency_kHz;
 
 // Tag IDs: so far, Lotek tags have 3 digit IDs.  But we want to be
 // able to distinguish among tags with the same Lotek ID but
-// (sufficiently) different burst intervals. 
+// (sufficiently) different burst intervals.   Also, we 
+// want to be able to give those tags a simple ID which a user in
+// the field can use to distinguish among tags.  Using the BI
+// is not helpful as it is hard to label a tag with that.
+// Instead, tags can be given additional indelible marks, and
+// the count of such marks distinguishes between tags with the
+// same freq and ID label in the hand.
+//
+// So normally, there are no marks and the tag ID is an integer.
+// e.g. 123 which can be thought of as 123.0
+// When a project has a second tag with ID 123, it is given an
+// indelible mark, and that tag ID is given as 123.1
+// If a third tag has ID 123, it is given two marks and called
+// 123.2, and so on.
 
-typedef int Lotek_Tag_ID;  // only 3 digits
+typedef float Lotek_Tag_ID;
 static const Lotek_Tag_ID BOGUS_LOTEK_TAG_ID = -1;
 
 // type representing an internal tag ID; each entry in the database
